requirements.txt
fastapi
uvicorn
numpy
requests
typesense
python-dotenv
pydantic
azure-ai-inference
azure-core
selenium
beautifulsoup4
requests
webdriver-manager
pillow
pytesseract
rank_bm25
numpy
faiss-cpu


